map-support added

+++ src/main/java/xml/test/Test.java
@@ -2,11 +2,11 @@ package xml.test;
 
 public class Test {
 
-    private int firstField;
+    private double firstField;
     TestSub classField;
     String stringField;
 
-    public Test(int a, String str, TestSub i) {
+    public Test(double a, String str, TestSub i) {
         this.firstField = a;
         this.stringField = str;
         classField = i;
